Update MyHealthContent component, add tests.
