+++ src/locales/en.json
@@ -364,6 +364,7 @@
     "guidance": "Guidance",
     "how_are_you_feeling": "How are you feeling today?",
     "log_symptoms": "Log symptoms",
+    "not_feeling_well": "Not Feeling Well?",
     "my_health": "My Health",
     "no_symptom_history": "No symptom history",
     "not_well": "Not well",
